fdf norm in work

+++ FDF/structs.h
@@ -160,7 +160,7 @@ typedef struct	s_data
 	int 		Q_bool;
 	int 		H_bool;
 	int 		F_bool;
-	int			disco;
+	float		disco;
 }				t_data;
 
 #endif
